Fix part 1
